backend/python-snippets/requirements.txt
torch>=2.2.0
transformers>=4.38.2
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
pgvector>=0.2.5
python-dotenv>=1.0.1
groq>=0.4.2
numpy>=1.26.4